requirements.txt
boto3==1.34.160
python-docx==1.1.2
PyMuPDF==1.24.9
loguru==0.7.2
setuptools==72.1.0
chromadb==0.5.3
langchain-chroma==0.1.0
